Reorganize schema files for better code organization
- Create base-schemas.ts for common schema components
- Create query-schemas.ts for query-related schemas
- Create resource-schemas.ts for Honeycomb resource schemas
- Maintain backward compatibility via schema.ts
- Remove duplicated collection-schemas.ts file
- Update all imports to use the new schema files
