+++ src/tools/list-boards.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { HoneycombAPI } from "../api/client.js";
 import { CollectionOptions } from "../types/api.js";
-import { ListBoardsSchema } from "../types/collection-schemas.js";
+import { ListBoardsSchema } from "../types/resource-schemas.js";
 import { createTool } from "../utils/tool-factory.js";
 import { handleCollection } from "../utils/collection.js";
